feat: switch to f64

+++ src/clip.rs
@@ -126,12 +126,12 @@ impl AudioClip {
         Ok(Self::from_raw(frames, sample_rate))
     }
 
-    pub fn sample(&self, position: f32) -> Option<Frame> {
-        let position = position * self.0.sample_rate as f32;
+    pub fn sample(&self, position: f64) -> Option<Frame> {
+        let position = position * self.0.sample_rate as f64;
         let actual_index = position as usize;
         if let Some(frame) = self.0.frames.get(actual_index) {
             let next_frame = self.0.frames.get(actual_index + 1).unwrap_or(frame);
-            Some(frame.interpolate(next_frame, position - actual_index as f32))
+            Some(frame.interpolate(next_frame, (position - actual_index as f64) as f32))
         } else {
             None
         }
@@ -152,7 +152,7 @@ impl AudioClip {
         self.0.frames.len()
     }
 
-    pub fn length(&self) -> f32 {
-        self.frame_count() as f32 / self.sample_rate() as f32
+    pub fn length(&self) -> f64 {
+        self.frame_count() as f64 / self.sample_rate() as f64
     }
 }
